requirements.txt
# Core dependencies
httpx>=0.24.0
pydantic>=2.0.0
PyYAML>=6.0.0
python-dotenv>=1.0.0

# Phase 1: Embedding-based detection (Tier 1 & 2)
sentence-transformers>=2.2.0
numpy>=1.24.0
torch>=2.0.0
faiss-cpu>=1.7.4  # Vector DB for semantic similarity search
qdrant-client>=1.7.0  # RAG-specific vector DB

# Phase 2: LLM Agent (Tier 3) - 100% Free
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.1.0
langchain-ollama>=0.1.0
structlog>=23.2.0
transformers>=4.35.0
accelerate>=0.24.0

# Phase 4: Production deployment
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
prometheus-client>=0.19.0
python-json-logger>=2.0.0
python-multipart>=0.0.6
redis>=5.0.0  # Distributed caching and rate limiting
slowapi>=0.1.9

# Phase 5: Authentication & Dashboard
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
streamlit>=1.28.0
plotly>=5.17.0
# Phase 6: Enterprise Observability (OpenTelemetry)
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation>=0.41b0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-httpx>=0.41b0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
